+++ sync-for-payables/Codat/Sync/Payables/CodatSyncPayables.cs
@@ -149,10 +149,10 @@ namespace Codat.Sync.Payables
         public SDKConfig SDKConfiguration { get; private set; }
 
         private const string _language = "csharp";
-        private const string _sdkVersion = "4.0.0";
-        private const string _sdkGenVersion = "2.413.0";
+        private const string _sdkVersion = "4.1.0";
+        private const string _sdkGenVersion = "2.429.0";
         private const string _openapiDocVersion = "3.0.0";
-        private const string _userAgent = "speakeasy-sdk/csharp 4.0.0 2.413.0 3.0.0 Codat.Sync.Payables";
+        private const string _userAgent = "speakeasy-sdk/csharp 4.1.0 2.429.0 3.0.0 Codat.Sync.Payables";
         private string _serverUrl = "";
         private int _serverIndex = 0;
         private ISpeakeasyHttpClient _client;